swinir_app/requirements.txt
streamlit
torch
torchvision
numpy
pandas
opencv-python
tifffile
scikit-image
lpips
psutil
pillow
pytorch-msssim